Replace Company with the resource implementation
