+++ lib/hubspot-ruby.rb
@@ -7,7 +7,6 @@ require 'hubspot/collection'
 require 'hubspot/paged_collection'
 require 'hubspot/properties'
 require 'hubspot/company'
-require 'hubspot/company2'
 require 'hubspot/company_properties'
 require 'hubspot/config'
 require 'hubspot/connection'
